tweaks from PR
